+++ NEWS.md
@@ -17,6 +17,10 @@
   selection arguments.
 * Can now also upgrade old MizerParams objects that do  not have a consistent
   initial_effort.
+* New function `validParams()` validates a MizerParams object and automatically
+  upgrades it with `upgradeParams()` if necessary.
+* Upgrade old params objects automatically when used in plot functions, rate
+  functions or in `project()`, #163.
 * Numerous small improvements to documentation.
 
 # mizer 2.0.3